video 기능 구현 완료

+++ findi_back/videos/views.py
@@ -12,7 +12,10 @@ def later_video_list_create(request):
     print(f'request.data {request.data}')
     if request.method == 'GET':
         videos = LaterVideo.objects.filter(user=request.user)
+        print(f'videos : {videos}')
         serializer = LaterVideoSerializer(videos, many=True)
+        print(f'serializer : {serializer.data}')
+
         return Response(serializer.data)
     
     if request.method == 'POST':
@@ -22,13 +25,20 @@ def later_video_list_create(request):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['DELETE'])
+@api_view(['GET', 'DELETE'])
 @permission_classes([IsAuthenticated])
-def later_video_delete(request, videoId):
+def later_video_detail_and_delete(request, videoId):
     try:
         video = LaterVideo.objects.get(user=request.user, videoId=videoId)
     except LaterVideo.DoesNotExist:
+        if request.method == 'GET':
+            return Response({'isSaved': False})
         return Response({'detail': 'Not found.'}, status=status.HTTP_404_NOT_FOUND)
 
-    video.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    if request.method == 'GET':
+        return Response({'isSaved': True})
+
+    if request.method == 'DELETE':
+        video.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
